+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Max power updated to 600 for Forsaken owners.
 * Bugfixes post-Forsaken launch.
 * Add Infamy rank to progress page.
+* Add "is:randomroll" search for items that have random rolls.
 * Remove "is:powermod" search.
 * Remove "basepower:" search.
 * Masterworks now have a gold border. Previously items with a power mod had a gold border, but there are no more power mods.
@@ -11,6 +12,7 @@
 * Review modes - say hello to Gambit (and goodbye to Trials, at least for a little while).
 * Added Etheric Spiral and Etheric Helix to the list of reputation items.
 * Fixed Year 1 weapons not having an elemental damage type.
+* Added "is:bow" search.
 
 # 4.68.3 (2018-09-03)